# backend/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "BDC API Documentation",
    "version": "1.0.7",
    "description": "BDC (Business Data Constructor) API 文档",
    "contact": {
      "name": "BDC Team"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3300",
      "description": "开发服务器"
    }
  ],
...
  "paths": {
    "/api/database-connections": {
      "post": {
        "tags": [
          "Database Connections"
        ],
        "operationId": "postDatabaseConnections",
        "summary": "创建数据库连接",
        "description": "创建新的数据库连接配置",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DatabaseConnectionCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "创建成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/DatabaseConnection"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误"
          },
          "500": {
            "description": "服务器错误"
          }
        }
      },
      "get": {
        "tags": [
          "Database Connections"
        ],
        "operationId": "getDatabaseConnections",
        "summary": "获取数据库连接列表",
        "description": "分页获取数据库连接列表，支持按类型、状态和是否激活进行筛选",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            },
            "description": "页码"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 10
            },
            "description": "每页数量"
          },
          {
            "in": "query",
            "name": "type",
            "schema": {
              "type": "string",
              "enum": [
                "POSTGRESQL",
                "MYSQL",
                "MONGODB",
                "SQLSERVER",
                "ORACLE"
              ]
            },
            "description": "数据库类型"
          },
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string",
              "enum": [
                "INACTIVE",
                "ACTIVE",
                "TESTING",
                "FAILED"
              ]
            },
            "description": "连接状态"
          },
          {
            "in": "query",
            "name": "isActive",
            "schema": {
              "type": "boolean"
            },
            "description": "是否激活"
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "items": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/DatabaseConnection"
                          }
                        },
                        "total": {
                          "type": "integer",
                          "example": 100
                        },
                        "page": {
                          "type": "integer",
                          "example": 1
                        },
                        "limit": {
                          "type": "integer",
                          "example": 10
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "服务器错误"
          }
        }
      }
    },
    "/api/database-connections/{id}/test": {
      "post": {
        "tags": [
          "Database Connections"
        ],
        "operationId": "postDatabaseConnectionsIdTest",
        "summary": "测试数据库连接",
        "description": "测试指定ID的数据库连接是否可用",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "数据库连接ID"
          }
        ],
        "responses": {
          "200": {
            "description": "测试成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "连接测试成功"
                    },
                    "data": {
                      "$ref": "#/components/schemas/DatabaseConnectionTestResult"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "数据库连接不存在"
          },
          "500": {
            "description": "服务器错误或连接测试失败"
          }
        }
      }
    },
    "/api/database-connections/{id}/tables": {
      "get": {
        "tags": [
          "Database Connections"
        ],
        "operationId": "getDatabaseConnectionsIdTables",
        "summary": "获取数据库表结构",
        "description": "获取指定数据库连接的表结构信息",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "数据库连接ID"
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/DatabaseTable"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "请先测试数据库连接"
          },
          "404": {
            "description": "数据库连接不存在"
          },
          "500": {
            "description": "服务器错误或获取表结构失败"
          }
        }
      }
    },
    "/api/database-connections/{id}": {
      "get": {
        "tags": [
          "Database Connections"
        ],
        "operationId": "getDatabaseConnectionsId",
        "summary": "获取单个数据库连接",
        "description": "根据ID获取数据库连接详情",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "数据库连接ID"
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/DatabaseConnection"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "数据库连接不存在"
          },
          "500": {
            "description": "服务器错误"
          }
        }
      },
      "put": {
        "tags": [
          "Database Connections"
        ],
        "operationId": "putDatabaseConnectionsId",
        "summary": "更新数据库连接",
        "description": "更新指定ID的数据库连接配置",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "数据库连接ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DatabaseConnectionUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "更新成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/DatabaseConnection"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误"
          },
          "404": {
            "description": "数据库连接不存在"
          },
          "500": {
            "description": "服务器错误"
          }
        }
      },
      "delete": {
        "tags": [
          "Database Connections"
        ],
        "operationId": "deleteDatabaseConnectionsId",
        "summary": "删除数据库连接",
        "description": "删除指定ID的数据库连接",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "数据库连接ID"
          }
        ],
        "responses": {
          "200": {
            "description": "删除成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "数据库连接已删除"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "无法删除正在使用的数据库连接"
          },
          "404": {
            "description": "数据库连接不存在"
          },
          "500": {
            "description": "服务器错误"
          }
        }
      }
    },
...
  },
  "components": {
...
    "schemas": {
...
      "DatabaseConnection": {
        "type": "object",
        "required": [
          "name",
          "type",
          "host",
          "port",
          "database",
          "username",
          "password"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "数据库连接ID"
          },
          "name": {
            "type": "string",
            "maxLength": 100,
            "description": "连接名称"
          },
          "description": {
            "type": "string",
            "maxLength": 100,
            "description": "连接描述"
          },
          "type": {
            "type": "string",
            "enum": [
              "postgresql",
              "mysql",
              "mongodb",
              "sqlserver",
              "oracle"
            ],
            "description": "数据库类型"
          },
          "host": {
            "type": "string",
            "maxLength": 255,
            "description": "主机地址"
          },
          "port": {
            "type": "integer",
            "minimum": 1,
            "maximum": 65535,
            "description": "端口号"
          },
          "database": {
            "type": "string",
            "maxLength": 100,
            "description": "数据库名称"
          },
          "username": {
            "type": "string",
            "maxLength": 100,
            "description": "用户名"
          },
          "password": {
            "type": "string",
            "maxLength": 255,
            "description": "密码（加密存储）"
          },
          "schema": {
            "type": "string",
            "maxLength": 100,
            "description": "Schema名称"
          },
          "sslConfig": {
            "type": "object",
            "description": "SSL配置",
            "properties": {
              "enabled": {
                "type": "boolean",
                "description": "是否启用SSL"
              },
              "verifyServerCert": {
                "type": "boolean",
                "description": "是否验证服务器证书"
              },
              "caCert": {
                "type": "string",
                "description": "CA证书内容"
              },
              "clientCert": {
                "type": "string",
                "description": "客户端证书内容"
              },
              "clientKey": {
                "type": "string",
                "description": "客户端密钥内容"
              }
            }
          },
          "poolConfig": {
            "type": "object",
            "description": "连接池配置",
            "properties": {
              "min": {
                "type": "integer",
                "description": "最小连接数"
              },
              "max": {
                "type": "integer",
                "description": "最大连接数"
              },
              "idleTimeoutMillis": {
                "type": "integer",
                "description": "空闲超时时间（毫秒）"
              },
              "connectionTimeoutMillis": {
                "type": "integer",
                "description": "连接超时时间（毫秒）"
              }
            }
          },
          "monitorConfig": {
            "type": "object",
            "description": "监控配置",
            "properties": {
              "enabled": {
                "type": "boolean",
                "description": "是否启用监控"
              },
              "checkInterval": {
                "type": "integer",
                "description": "检查间隔（秒）"
              },
              "metrics": {
                "type": "array",
                "items": {
                  "type": "string"
                },
                "description": "监控指标列表"
              },
              "alertThresholds": {
                "type": "object",
                "description": "告警阈值",
                "properties": {
                  "maxConnections": {
                    "type": "integer",
                    "description": "最大连接数"
                  },
                  "maxQueryTime": {
                    "type": "integer",
                    "description": "最大查询时间"
                  },
                  "maxErrorRate": {
                    "type": "number",
                    "description": "最大错误率"
                  }
                }
              }
            }
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "inactive",
              "testing",
              "failed",
              "maintenance"
            ],
            "description": "连接状态"
          },
          "allowRemote": {
            "type": "boolean",
            "description": "是否支持远程连接"
          },
          "allowedIps": {
            "type": "string",
            "maxLength": 255,
            "description": "允许的IP地址列表，多个用逗号分隔"
          },
          "lastTestAt": {
            "type": "string",
            "format": "date-time",
            "description": "最后测试连接时间"
          },
          "lastTestSuccess": {
            "type": "boolean",
            "description": "最后测试是否成功"
          },
          "lastTestError": {
            "type": "string",
            "maxLength": 500,
            "description": "最后测试错误信息"
          },
          "stats": {
            "type": "object",
            "description": "连接统计信息",
            "properties": {
              "totalConnections": {
                "type": "integer",
                "description": "总连接数"
              },
              "activeConnections": {
                "type": "integer",
                "description": "活跃连接数"
              },
              "failedConnections": {
                "type": "integer",
                "description": "失败连接数"
              },
              "lastErrorAt": {
                "type": "string",
                "format": "date-time",
                "description": "最后错误时间"
              },
              "lastError": {
                "type": "string",
                "description": "最后错误信息"
              },
              "avgQueryTime": {
                "type": "number",
                "description": "平均查询时间"
              },
              "maxQueryTime": {
                "type": "number",
                "description": "最大查询时间"
              }
            }
          },
          "isActive": {
            "type": "boolean",
            "description": "是否启用"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "创建时间"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "更新时间"
          }
        }
      },
      "DatabaseConnectionCreate": {
        "type": "object",
        "required": [
          "name",
          "type",
          "host",
          "port",
          "database",
          "username",
          "password"
        ],
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 100,
            "description": "连接名称"
          },
          "description": {
            "type": "string",
            "maxLength": 100,
            "description": "连接描述"
          },
          "type": {
            "type": "string",
            "enum": [
              "postgresql",
              "mysql",
              "mongodb",
              "sqlserver",
              "oracle"
            ],
            "description": "数据库类型"
          },
          "host": {
            "type": "string",
            "maxLength": 255,
            "description": "主机地址"
          },
          "port": {
            "type": "integer",
            "minimum": 1,
            "maximum": 65535,
            "description": "端口号"
          },
          "database": {
            "type": "string",
            "maxLength": 100,
            "description": "数据库名称"
          },
          "username": {
            "type": "string",
            "maxLength": 100,
            "description": "用户名"
          },
          "password": {
            "type": "string",
            "maxLength": 255,
            "description": "密码"
          },
          "schema": {
            "type": "string",
            "maxLength": 100,
            "description": "Schema名称"
          },
          "sslConfig": {
            "$ref": "#/components/schemas/DatabaseConnection/properties/sslConfig"
          },
          "poolConfig": {
            "$ref": "#/components/schemas/DatabaseConnection/properties/poolConfig"
          },
          "monitorConfig": {
            "$ref": "#/components/schemas/DatabaseConnection/properties/monitorConfig"
          },
          "allowRemote": {
            "type": "boolean",
            "description": "是否支持远程连接"
          },
          "allowedIps": {
            "type": "string",
            "maxLength": 255,
            "description": "允许的IP地址列表，多个用逗号分隔"
          }
        }
      },
      "DatabaseConnectionUpdate": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 100,
            "description": "连接名称"
          },
          "description": {
            "type": "string",
            "maxLength": 100,
            "description": "连接描述"
          },
          "host": {
            "type": "string",
            "maxLength": 255,
            "description": "主机地址"
          },
          "port": {
            "type": "integer",
            "minimum": 1,
            "maximum": 65535,
            "description": "端口号"
          },
          "database": {
            "type": "string",
            "maxLength": 100,
            "description": "数据库名称"
          },
          "username": {
            "type": "string",
            "maxLength": 100,
            "description": "用户名"
          },
          "password": {
            "type": "string",
            "maxLength": 255,
            "description": "密码"
          },
          "schema": {
            "type": "string",
            "maxLength": 100,
            "description": "Schema名称"
          },
          "sslConfig": {
            "$ref": "#/components/schemas/DatabaseConnection/properties/sslConfig"
          },
          "poolConfig": {
            "$ref": "#/components/schemas/DatabaseConnection/properties/poolConfig"
          },
          "monitorConfig": {
            "$ref": "#/components/schemas/DatabaseConnection/properties/monitorConfig"
          },
          "allowRemote": {
            "type": "boolean",
            "description": "是否支持远程连接"
          },
          "allowedIps": {
            "type": "string",
            "maxLength": 255,
            "description": "允许的IP地址列表，多个用逗号分隔"
          },
          "isActive": {
            "type": "boolean",
            "description": "是否启用"
          }
        }
      },
...
      "DatabaseConnectionTestResult": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "测试是否成功"
          },
          "message": {
            "type": "string",
            "description": "测试结果消息"
          },
          "connectionInfo": {
            "type": "object",
            "description": "连接信息",
            "properties": {
              "version": {
                "type": "string",
                "description": "数据库版本"
              },
              "serverTime": {
                "type": "string",
                "format": "date-time",
                "description": "服务器时间"
              },
              "maxConnections": {
                "type": "integer",
                "description": "最大连接数"
              },
              "currentConnections": {
                "type": "integer",
                "description": "当前连接数"
              }
            }
          }
        }
      },
      "DatabaseTable": {
        "type": "object",
        "required": [
          "tableName",
          "schema",
          "columns"
        ],
        "properties": {
          "tableName": {
            "type": "string",
            "description": "表名"
          },
          "schema": {
            "type": "string",
            "description": "Schema名称"
          },
          "description": {
            "type": "string",
            "description": "表描述"
          },
          "tableType": {
            "type": "string",
            "description": "表类型"
          },
          "rowCount": {
            "type": "integer",
            "description": "行数"
          },
          "size": {
            "type": "integer",
            "description": "表大小（字节）"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "创建时间"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "最后修改时间"
          },
          "columns": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DatabaseColumn"
            },
            "description": "字段列表"
          },
          "indexes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DatabaseIndex"
            },
            "description": "索引列表"
          },
          "foreignKeys": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DatabaseForeignKey"
            },
            "description": "外键列表"
          }
        }
      },
      "DatabaseColumn": {
        "type": "object",
        "required": [
          "name",
          "type",
          "nullable",
          "primaryKey"
        ],
        "properties": {
          "name": {
            "type": "string",
            "description": "字段名"
          },
          "type": {
            "type": "string",
            "description": "数据类型"
          },
          "nullable": {
            "type": "boolean",
            "description": "是否为空"
          },
          "primaryKey": {
            "type": "boolean",
            "description": "是否为主键"
          },
          "defaultValue": {
            "type": "string",
            "description": "默认值"
          },
          "description": {
            "type": "string",
            "description": "字段描述"
          },
          "length": {
            "type": "integer",
            "description": "字段长度"
          },
          "precision": {
            "type": "integer",
            "description": "精度（用于decimal类型）"
          },
          "scale": {
            "type": "integer",
            "description": "小数位数（用于decimal类型）"
          },
          "autoIncrement": {
            "type": "boolean",
            "description": "是否自增"
          },
          "ordinalPosition": {
            "type": "integer",
            "description": "字段位置"
          },
          "characterSet": {
            "type": "string",
            "description": "字符集"
          },
          "collation": {
            "type": "string",
            "description": "排序规则"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "最后修改时间"
          }
        }
      },
      "DatabaseIndex": {
        "type": "object",
        "required": [
          "name",
          "type",
          "columns"
        ],
        "properties": {
          "name": {
            "type": "string",
            "description": "索引名"
          },
          "type": {
            "type": "string",
            "enum": [
              "PRIMARY",
              "UNIQUE",
              "INDEX"
            ],
            "description": "索引类型"
          },
          "columns": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "索引字段列表"
          },
          "description": {
            "type": "string",
            "description": "索引描述"
          }
        }
      },
...
    }
  }
}